Update model.py
modified network to be implemented like paper's implementations

In the original paper, they mentioned.

"For the ESPCN, we set l = 3, (f1, n1) = (5, 64), (f2, n2) = (3, 32) and f3 = 3 in our evaluations. The choice of the parameter is inspired by SRCNN’s 3 layer 9-5-5 model and the equations in Sec. 2.2."

So, I removed one layer and adjusted the number of filters.